Completed Section 2.6

+++ functions/API/todos.js
@@ -1,9 +1,9 @@
 const { db } = require("../util/admin");
 
 exports.getAllTodos = (request, response) => {
-    console.log("OK")
     db
         .collection("todos")
+        .where("username", "==", request.user.username)
         .orderBy("createdAt", "desc")
         .get()
         .then((data) => {
@@ -24,6 +24,27 @@ exports.getAllTodos = (request, response) => {
         });
 };
 
+exports.getOneTodo = (request, response) => {
+    const document = db.doc(`/todos/${request.params.todoId}`);
+    document
+        .get()
+        .then((doc) => {
+            if (!doc.exists) {
+                return response.status(404).json({ error: "Todo not found"});
+            }
+            return response.json({
+                todoId: doc.id,
+                title: doc.data().title,
+                body: doc.data().body,
+                createdAt: doc.data().createdAt,
+            });
+        })
+        .catch((err) => {
+            console.error(err);
+            return response.status(500).json( {error: err.code })
+        });
+}
+
 exports.postOneTodo = (request, response) => {
     if (request.body.body.trim() === "") {
         return response.status(400).json({ body: "Must not be empty" });
@@ -35,7 +56,8 @@ exports.postOneTodo = (request, response) => {
     const newTodoItem = {
         title: request.body.title,
         body: request.body.body,
-        createdAt: new Date().toISOString()
+        createdAt: new Date().toISOString(),
+        username: request.user.username
     }
     db
         .collection("todos")
@@ -58,6 +80,9 @@ exports.deleteTodo = (request, response) => {
         .then((doc) => {
             if (!doc.exists) {
                 return response.status(404).json({ error: "Todo not found" });
+            }
+            if (doc.data().username !== request.user.username) {
+                return response.status(403).json({ error: "Unauthorised" });
             }
             return document.delete();
         })
